refactor(react-centra-checkout): restructure files
By doing so, we align with the other packages in the accelerator.
